Location improvements (#30)

+++ scripts/mysql/100-mysql_init.sql
@@ -49,8 +49,8 @@ CREATE TABLE Locations(
   user_id INT NOT NULL,
   latitude FLOAT NOT NULL,
   longitude FLOAT NOT NULL,
-  accuracy FLOAT NOT NULL,
-  time_stamp FLOAT NOT NULL,
+  accuracy INT NOT NULL,
+  time_stamp BIGINT NOT NULL,
   PRIMARY KEY (user_id),
   CONSTRAINT FK_User FOREIGN KEY (user_id) REFERENCES User(id)
 );
